src/requirements.txt
sugar==0.9.2
rerunfailures==5.0
ordering==0.6
metadata==1.8.0
instafail==0.4.1
datafiles==2.0
assume==1.2.2
datatest==0.9.3
allure-pytest==2.6.2